Geocoding / Geolocation
Building on @vbdjames work with the Smarty Streets API, split
responsibilities for geocoding and geolocation into two separate
directives.  In modal, when user uses one option or the other,
appropriate calls are made and the modal is closed on completion.
Results appear in the Geosearch scope as usual.  Location button in
upper right reopens modal.